+++ vm/src/hint_processor/builtin_hint_processor/ec_utils.rs
@@ -225,7 +225,7 @@ mod tests {
     use crate::any_box;
     use crate::hint_processor::builtin_hint_processor::builtin_hint_processor_definition::BuiltinHintProcessor;
     use crate::hint_processor::builtin_hint_processor::builtin_hint_processor_definition::HintProcessorData;
-    use crate::hint_processor::hint_processor_definition::HintProcessor;
+    use crate::hint_processor::hint_processor_definition::HintProcessorLogic;
     use crate::relocatable;
     use crate::types::exec_scope::ExecutionScopes;
     use crate::types::relocatable::Relocatable;
@@ -233,7 +233,6 @@ mod tests {
 
     use crate::hint_processor::builtin_hint_processor::hint_code;
     use crate::utils::test_utils::*;
-    use crate::vm::runners::cairo_runner::RunResources;
     use assert_matches::assert_matches;
 
     use super::*;