+++ vm/src/hint_processor/builtin_hint_processor/uint384.rs
@@ -230,7 +230,6 @@ pub fn sub_reduced_a_and_reduced_b(
 mod tests {
     use super::*;
     use crate::hint_processor::builtin_hint_processor::hint_code;
-    use crate::vm::runners::cairo_runner::RunResources;
     use core::ops::Shl;
 
     use crate::{
@@ -239,7 +238,7 @@ mod tests {
             builtin_hint_processor::builtin_hint_processor_definition::{
                 BuiltinHintProcessor, HintProcessorData,
             },
-            hint_processor_definition::HintProcessor,
+            hint_processor_definition::HintProcessorLogic,
         },
         types::{
             exec_scope::ExecutionScopes,