+++ vm/src/vm/vm_core.rs
@@ -29,7 +29,6 @@ use num_traits::{ToPrimitive, Zero};
 
 use super::errors::trace_errors::TraceError;
 use super::runners::builtin_runner::OUTPUT_BUILTIN_NAME;
-use super::runners::cairo_runner::RunResources;
 
 const MAX_TRACEBACK_ENTRIES: u32 = 20;
 
@@ -449,12 +448,11 @@ impl VirtualMachine {
         exec_scopes: &mut ExecutionScopes,
         hint_data_dictionary: &HashMap<usize, Vec<Box<dyn Any>>>,
         constants: &HashMap<String, Felt252>,
-        run_resources: &mut RunResources,
     ) -> Result<(), VirtualMachineError> {
         if let Some(hint_list) = hint_data_dictionary.get(&self.run_context.pc.offset) {
             for (hint_index, hint_data) in hint_list.iter().enumerate() {
                 hint_executor
-                    .execute_hint(self, exec_scopes, hint_data, constants, run_resources)
+                    .execute_hint(self, exec_scopes, hint_data, constants)
                     .map_err(|err| VirtualMachineError::Hint(Box::new((hint_index, err))))?
             }
         }
@@ -488,15 +486,8 @@ impl VirtualMachine {
         exec_scopes: &mut ExecutionScopes,
         hint_data_dictionary: &HashMap<usize, Vec<Box<dyn Any>>>,
         constants: &HashMap<String, Felt252>,
-        run_resources: &mut RunResources,
     ) -> Result<(), VirtualMachineError> {
-        self.step_hint(
-            hint_executor,
-            exec_scopes,
-            hint_data_dictionary,
-            constants,
-            run_resources,
-        )?;
+        self.step_hint(hint_executor, exec_scopes, hint_data_dictionary, constants)?;
 
         #[cfg(feature = "hooks")]
         self.execute_pre_step_instruction(
@@ -2669,7 +2660,6 @@ mod tests {
                 exec_scopes_ref!(),
                 &HashMap::new(),
                 &HashMap::new(),
-                &mut RunResources::default(),
             ),
             Ok(())
         );
@@ -2899,7 +2889,6 @@ mod tests {
                 exec_scopes_ref!(),
                 &HashMap::new(),
                 &HashMap::new(),
-                &mut RunResources::default(),
             ),
             Ok(())
         );
@@ -2982,7 +2971,6 @@ mod tests {
                     exec_scopes_ref!(),
                     &HashMap::new(),
                     &HashMap::new(),
-                    &mut RunResources::default(),
                 ),
                 Ok(())
             );
@@ -3079,7 +3067,6 @@ mod tests {
                 exec_scopes_ref!(),
                 &HashMap::new(),
                 &HashMap::new(),
-                &mut RunResources::default(),
             ),
             Ok(())
         );
@@ -3101,7 +3088,6 @@ mod tests {
                 exec_scopes_ref!(),
                 &HashMap::new(),
                 &HashMap::new(),
-                &mut RunResources::default(),
             ),
             Ok(())
         );
@@ -3124,7 +3110,6 @@ mod tests {
                 exec_scopes_ref!(),
                 &HashMap::new(),
                 &HashMap::new(),
-                &mut RunResources::default(),
             ),
             Ok(())
         );
@@ -3690,7 +3675,6 @@ mod tests {
                     exec_scopes_ref!(),
                     &hint_data_dictionary,
                     &HashMap::new(),
-                    &mut RunResources::default(),
                 ),
                 Ok(())
             );
@@ -4167,12 +4151,7 @@ mod tests {
 
         let end = cairo_runner.initialize(&mut vm).unwrap();
         assert!(cairo_runner
-            .run_until_pc(
-                end,
-                &mut RunResources::default(),
-                &mut vm,
-                &mut hint_processor
-            )
+            .run_until_pc(end, &mut vm, &mut hint_processor)
             .is_err());
         let expected_traceback = vec![
             (Relocatable::from((1, 3)), Relocatable::from((0, 97))),
@@ -4197,12 +4176,7 @@ mod tests {
 
         let end = cairo_runner.initialize(&mut vm).unwrap();
         assert!(cairo_runner
-            .run_until_pc(
-                end,
-                &mut RunResources::default(),
-                &mut vm,
-                &mut hint_processor
-            )
+            .run_until_pc(end, &mut vm, &mut hint_processor)
             .is_err());
         let expected_traceback = vec![(Relocatable::from((1, 2)), Relocatable::from((0, 34)))];
         assert_eq!(vm.get_traceback_entries(), expected_traceback);
@@ -4291,12 +4265,7 @@ mod tests {
                 .unwrap();
 
             assert!(cairo_runner
-                .run_until_pc(
-                    end,
-                    &mut RunResources::default(),
-                    &mut virtual_machine_from_builder,
-                    &mut hint_processor
-                )
+                .run_until_pc(end, &mut virtual_machine_from_builder, &mut hint_processor)
                 .is_err());
         }
     }